Moved build/validator.go to client/buildconfigvalidator
