+++ build_signaler/signaler.go
@@ -8,15 +8,15 @@ import (
 	"github.com/shankj3/go-til/log"
 	"github.com/shankj3/go-til/nsqpb"
 
-	"github.com/level11consulting/ocelot/build"
 	"github.com/level11consulting/ocelot/build/helpers/messageservice"
 	stringbuilder "github.com/level11consulting/ocelot/build/helpers/stringbuilder/accountrepo"
+	"github.com/level11consulting/ocelot/client/buildconfigvalidator"
 	"github.com/level11consulting/ocelot/client/runtime"
 	"github.com/level11consulting/ocelot/server/config"
 	"github.com/level11consulting/ocelot/storage"
 )
 
-func NewSignaler(RC config.CVRemoteConfig, dese *deserialize.Deserializer, producer nsqpb.Producer, ocyValidator *build.OcelotValidator, store storage.OcelotStorage) *Signaler {
+func NewSignaler(RC config.CVRemoteConfig, dese *deserialize.Deserializer, producer nsqpb.Producer, ocyValidator *buildconfigvalidator.OcelotValidator, store storage.OcelotStorage) *Signaler {
 	return &Signaler{
 		RC:           RC,
 		Deserializer: dese,
@@ -30,7 +30,7 @@ type Signaler struct {
 	RC config.CVRemoteConfig
 	*deserialize.Deserializer
 	Producer     nsqpb.Producer
-	OcyValidator *build.OcelotValidator
+	OcyValidator *buildconfigvalidator.OcelotValidator
 	Store        storage.OcelotStorage
 }
 
@@ -63,7 +63,7 @@ func (s *Signaler) QueueAndStore(task *pb.WerkerTask) error {
 	alreadyBuilding, err := runtime.CheckBuildInConsul(s.RC.GetConsul(), task.CheckoutHash)
 	if alreadyBuilding {
 		log.Log().Info("kindly refusing to add to queue because this hash is already building")
-		return build.NoViability("this hash is already building in ocelot, therefore not adding to queue")
+		return buildconfigvalidator.NoViability("this hash is already building in ocelot, therefore not adding to queue")
 	}
 	//get vcs cred to attach the database id to the build summary
 	vcsCred, err := config.GetVcsCreds(s.Store, task.FullName, s.RC, task.VcsType)
